miProyecto/requirements.txt
asgiref==3.7.2
Brotli==1.1.0
click==8.1.7
dj-database-url==2.2.0
Django==4.2.7
django-filter==24.1
djangorestframework==3.14.0
gunicorn==22.0.0
h11==0.14.0
Markdown==3.5.2
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
uvicorn==0.30.0
whitenoise==6.6.0
